feedback popup

+++ src/common/convertDate.js
@@ -1,9 +1,9 @@
 export const formatDatePublic = (dateString) => {
-  const options = { year: "numeric", month: "long" }
+  const options = { year: "numeric", month: "short" }
   return new Date(dateString).toLocaleDateString(undefined, options)
 }
 
 export const formatDatePrivate = (dateString) => {
-  const options = { year: "numeric", month: "long", day: "numeric" }
+  const options = { year: "numeric", month: "short", day: "numeric" }
   return new Date(dateString).toLocaleDateString(undefined, options)
 }
